backend/zzz_fastapi_backup/2.3.2 JWT 双 Token Demo/app/db/database.py
import os

from dotenv import load_dotenv
from sqlalchemy.ext.asyncio import AsyncSession, create_async_engine
from sqlalchemy.orm import declarative_base, sessionmaker

load_dotenv()

DATABASE_URL = os.getenv("ASYNC_DATABASE_URL")
engine = create_async_engine(DATABASE_URL, future=True, echo=False)
AsyncSessionLocal = sessionmaker(engine, class_=AsyncSession, expire_on_commit=False)
Base = declarative_base()


# Dependency to get async session
async def get_db():
    async with AsyncSessionLocal() as session:
        yield session
